Fixed bug where handing in nil for these parameters could cut properties dictionary
To better support the possibility of someone handing in nil as a
"title", "isbn" or "author", I recommend explicitly setting the keys and
values on the property dictionaries rather than using
"dictionaryWithObjectsAndKeys:".

I discovered this bug when I was handing in a nil ISBN since I didn't
have it when creating a book record. Since it was passed directly to
"dictionaryWithObjectsAndKeys:", then that method thought there were no
more keys or values and stopped processing them.